Add StoreAdapter that combines the in-memory and Redis approaches
Since the in-memory storage adapter has the drawback of each node populating it's own cache and the Redis storage adapter requires remote calls over the network, this `RedisWithLocalCacheStoreAdapter` takes a hybrid approach that keeps an in-memory cache fresh while delegating cache misses to the Redis adapter for in-memory population.  This allows us to force clients into the "new query path" less frequently and avoids unnecessary remote network calls.